Fix warnings: default arg is mutable
This was mostly `decomposition_kwargs` and `grid_boundaries`, in various
methods,which were mutable dicts and lists respectively. Replace the lists with
tuples and replace the dicts with `None`. These `None` will then be replaced
with an empty dict (like the prevois default) when deconstructed
(i.e. like `**kwargs`).

Additionally, reformat:
* autopep8 datasets/positional_encoding.py
* autopep8 layers/spherical_convolution.py
* autopep8 datasets/spherical_swe.py